revert back to i3-contained deprecation warning
